Write ptxt_distance()
If two strings hashes to different values, they are regarded as
different strings. And if they hashes to the same values, they are
regarded as identical string.

TODO:
Write str_hash().
Keep the shortest path in dp table.
I dont want 1234(EOF) and 1234(\n) to be recognized as different text.